Release 0.13.21: Extract GitCredentials + ContainerFilePush from ProjectProvisioner
ProjectProvisioner was a 1530-line god-file mixing git-credential resolution
and the incus-file-push idiom with the provisioning pipeline. Extract two
cohesive, independently-tested utilities:

- GitCredentials: buildCredentialStore / extractHttpsHosts / extractSshHosts /
  resolveTokenForHost / singleTokenMap / resolveHostPath, plus the private
  host-safety validation (requireSafeHost/invalidPort/HOST_PATTERN). These were
  static methods that ProjectApplier and four command classes already reached
  into across the ProjectProvisioner boundary.
- ContainerFilePush: the "stage to host temp file, incus file push, delete in
  finally" idiom that was duplicated in ProjectProvisioner.pushFileToContainer
  and ProjectApplier.pushFile, parameterized by the push flags so both the
  root-owned (--mode) and dev-owned (--uid/--gid) call sites share it.

Behavior-preserving: messages, the HOST_PATTERN regex, and the push command
shapes are unchanged; ProjectApplier still does its dev-user mkdir before push.
Credential/path unit tests moved to GitCredentialsTest (18); ContainerFilePush
covered by a new test (4). ProjectProvisioner drops 176 lines.

Full suite green (core 862, harness 120, infra 1424); JaCoCo gates met.
