+++ sail-core/pom.xml
@@ -7,7 +7,7 @@
     <parent>
         <groupId>ai.singlr</groupId>
         <artifactId>sail</artifactId>
-        <version>0.13.20</version>
+        <version>0.13.21</version>
     </parent>
 
     <artifactId>sail-core</artifactId>
