+++ sail-infra/src/test/java/ai/singlr/sail/engine/ProjectProvisionerTest.java
@@ -1072,26 +1072,6 @@ class ProjectProvisionerTest {
   }
 
   @Test
-  void resolveHostPathExpandsTildeWithSudoUser() {
-    var original = System.getenv("SUDO_USER");
-    var path = ProjectProvisioner.resolveHostPath("~/.ssh/id_ed25519");
-    if (original != null) {
-      assertEquals(Path.of("/home", original, ".ssh/id_ed25519"), path);
-    } else {
-      assertEquals(
-          Path.of(System.getProperty("user.home"), ".ssh/id_ed25519"),
-          path,
-          "Should fall back to user.home when SUDO_USER is not set");
-    }
-  }
-
-  @Test
-  void resolveHostPathReturnsAbsolutePathUnchanged() {
-    var path = ProjectProvisioner.resolveHostPath("/tmp/my-key");
-    assertEquals(Path.of("/tmp/my-key"), path);
-  }
-
-  @Test
   void gitSshKeyPushedIntoContainer() throws Exception {
     var keyFile = tempDir.resolve("id_ed25519");
     Files.writeString(
@@ -1133,144 +1113,6 @@ class ProjectProvisionerTest {
             IOException.class, () -> provisioner.provision(config, hostYaml(), null, null));
     assertTrue(ex.getMessage().contains("SSH key not found"));
     assertTrue(ex.getMessage().contains("git.ssh_key"));
-  }
-
-  @Test
-  void buildCredentialStoreEmptyWhenNoRepos() {
-    var creds = ProjectProvisioner.buildCredentialStore(Map.of("*", "tok123"), null);
-    assertEquals("", creds);
-  }
-
-  @Test
-  void buildCredentialStoreIncludesAllHttpsHosts() {
-    var repos =
-        List.of(
-            new SailYaml.Repo("https://gitlab.com/org/repo.git", "repo", null),
-            new SailYaml.Repo("https://bitbucket.org/team/app.git", "app", null),
-            new SailYaml.Repo("git@github.com:user/other.git", "other", null));
-    var creds = ProjectProvisioner.buildCredentialStore(Map.of("*", "tok123"), repos);
-    assertTrue(creds.contains("https://oauth2:tok123@gitlab.com\n"));
-    assertTrue(creds.contains("https://oauth2:tok123@bitbucket.org\n"));
-    assertFalse(creds.contains("git@"), "SSH URLs should not appear in credential store");
-  }
-
-  @Test
-  void buildCredentialStoreUsesPerHostTokens() {
-    var repos =
-        List.of(
-            new SailYaml.Repo("https://github.com/org/app.git", "app", null),
-            new SailYaml.Repo("https://gitlab.com/org/repo.git", "repo", null));
-    var tokens = Map.of("github.com", "ghp_xxx", "gitlab.com", "glpat_yyy");
-    var creds = ProjectProvisioner.buildCredentialStore(tokens, repos);
-    assertTrue(creds.contains("https://oauth2:ghp_xxx@github.com\n"));
-    assertTrue(creds.contains("https://oauth2:glpat_yyy@gitlab.com\n"));
-  }
-
-  @Test
-  void buildCredentialStoreDeduplicatesHosts() {
-    var repos =
-        List.of(
-            new SailYaml.Repo("https://gitlab.com/org/a.git", "a", null),
-            new SailYaml.Repo("https://gitlab.com/org/b.git", "b", null));
-    var creds = ProjectProvisioner.buildCredentialStore(Map.of("*", "tok123"), repos);
-    var lines = creds.strip().split("\n");
-    assertEquals(1, lines.length, "One gitlab.com entry (deduplicated)");
-  }
-
-  @Test
-  void resolveTokenForHostChecksExplicitFirst() {
-    var tokens = Map.of("gitlab.com", "explicit-tok", "*", "wildcard-tok");
-    var result = ProjectProvisioner.resolveTokenForHost("gitlab.com", tokens);
-    assertEquals("explicit-tok", result);
-  }
-
-  @Test
-  void resolveTokenForHostFallsBackToWildcard() {
-    var tokens = Map.of("*", "wildcard-tok");
-    var result = ProjectProvisioner.resolveTokenForHost("gitea.example.com", tokens);
-    assertEquals("wildcard-tok", result);
-  }
-
-  @Test
-  void resolveTokenForHostReturnsNullWhenNoMatch() {
-    var result = ProjectProvisioner.resolveTokenForHost("gitea.self-hosted.com", Map.of());
-    assertNull(result);
-  }
-
-  @Test
-  void buildCredentialStoreSkipsHostsWithNoToken() {
-    var repos =
-        List.of(
-            new SailYaml.Repo("https://gitlab.com/org/repo.git", "repo", null),
-            new SailYaml.Repo("https://github.com/org/app.git", "app", null));
-    var creds = ProjectProvisioner.buildCredentialStore(null, repos);
-    for (var line : creds.lines().toList()) {
-      assertFalse(line.contains("null"), "Should not include null token entries");
-    }
-  }
-
-  @Test
-  void buildCredentialStoreWorksWithEmptyMap() {
-    var repos = List.of(new SailYaml.Repo("https://github.com/org/app.git", "app", null));
-    var creds = ProjectProvisioner.buildCredentialStore(Map.of(), repos);
-    assertNotNull(creds);
-  }
-
-  @Test
-  void singleTokenMapWrapsToken() {
-    var result = ProjectProvisioner.singleTokenMap("my-token");
-    assertEquals(Map.of("*", "my-token"), result);
-  }
-
-  @Test
-  void singleTokenMapReturnsEmptyForNull() {
-    var result = ProjectProvisioner.singleTokenMap(null);
-    assertTrue(result.isEmpty());
-  }
-
-  @Test
-  void extractHttpsHostsFromMixedRepos() {
-    var repos =
-        List.of(
-            new SailYaml.Repo("https://github.com/org/app.git", "app", null),
-            new SailYaml.Repo("https://gitlab.com/org/repo.git", "repo", null),
-            new SailYaml.Repo("git@bitbucket.org:team/lib.git", "lib", null));
-    var hosts = ProjectProvisioner.extractHttpsHosts(repos);
-    assertEquals(List.of("github.com", "gitlab.com"), hosts);
-  }
-
-  @Test
-  void extractSshHostsIncludesDefaults() {
-    var hosts = ProjectProvisioner.extractSshHosts(null);
-    assertTrue(hosts.contains("github.com"));
-    assertTrue(hosts.contains("gitlab.com"));
-  }
-
-  @Test
-  void extractSshHostsFromMixedUrls() {
-    var repos =
-        List.of(
-            new SailYaml.Repo("git@bitbucket.org:team/app.git", "app", null),
-            new SailYaml.Repo("https://gitlab.example.com/org/repo.git", "repo", null));
-    var hosts = ProjectProvisioner.extractSshHosts(repos);
-    assertTrue(hosts.contains("github.com"), "Default github.com");
-    assertTrue(hosts.contains("gitlab.com"), "Default gitlab.com");
-    assertTrue(hosts.contains("bitbucket.org"), "SSH URL host");
-    assertTrue(hosts.contains("gitlab.example.com"), "HTTPS URL host");
-  }
-
-  @Test
-  void extractSshHostsRejectsUnsafeHost() {
-    var repos = List.of(new SailYaml.Repo("https://github.com;touch/tmp/pwned", "app", null));
-
-    assertThrows(IllegalArgumentException.class, () -> ProjectProvisioner.extractSshHosts(repos));
-  }
-
-  @Test
-  void extractHttpsHostsRejectsUnsafeHost() {
-    var repos = List.of(new SailYaml.Repo("https://github.com;touch/tmp/pwned", "app", null));
-
-    assertThrows(IllegalArgumentException.class, () -> ProjectProvisioner.extractHttpsHosts(repos));
   }
 
   private static SailYaml configWithGitSshKey(String keyPath) {
